backend/requirements.txt
Authlib==1.3.2
blinker==1.9.0
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
fuzzywuzzy==0.18.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
Levenshtein==0.26.1
MarkupSafe==3.0.2
matplotlib==3.10.0
moto==5.0.23
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
responses==0.25.3
s3transfer==0.10.4
six==1.17.0
tenacity==9.0.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
xmltodict==0.14.2
